implement unified ETLS and use it for all host backends
