+++ include/oneapi/dpl/pstl/omp/util.h
@@ -155,87 +155,33 @@ __process_chunk(const __chunk_metrics& __metrics, _Iterator __base, _Index __chu
     __f(__first, __last);
 }
 
-namespace __detail
+struct __get_num_threads
 {
-
-template <typename _ValueType, typename... _Args>
-struct __enumerable_thread_local_storage
-{
-    template <typename... _LocalArgs>
-    __enumerable_thread_local_storage(_LocalArgs&&... __args)
-        : __num_elements(0), __args(std::forward<_LocalArgs>(__args)...)
-    {
-        std::size_t __num_threads = omp_in_parallel() ? omp_get_num_threads() : omp_get_max_threads();
-        __thread_specific_storage.resize(__num_threads);
-    }
-
-    // Note: size should not be used concurrently with parallel loops which may instantiate storage objects, as it may
-    // not return an accurate count of instantiated storage objects in lockstep with the number allocated and stored.
-    // This is because the count is not atomic with the allocation and storage of the storage objects.
     std::size_t
-    size() const
+    operator()()
     {
-        // only count storage which has been instantiated
-        return __num_elements.load();
+        return omp_in_parallel() ? omp_get_num_threads() : omp_get_max_threads();
     }
-
-    // Note: get_with_id should not be used concurrently with parallel loops which may instantiate storage objects,
-    // as its operation may provide an out of date view of the stored objects based on the timing new object creation
-    // and incrementing of the size.
-    // TODO: Consider replacing this access with a visitor pattern.
-    _ValueType&
-    get_with_id(std::size_t __i)
-    {
-        assert(__i < size());
-
-        std::size_t __j = 0;
-
-        if (size() == __thread_specific_storage.size())
-        {
-            return *__thread_specific_storage[__i];
-        }
-
-        for (std::size_t __count = 0; __j < __thread_specific_storage.size() && __count <= __i; ++__j)
-        {
-            // Only include storage from threads which have instantiated a storage object
-            if (__thread_specific_storage[__j])
-            {
-                ++__count;
-            }
-        }
-        // Need to back up one once we have found a valid storage object
-        return *__thread_specific_storage[__j - 1];
-    }
-
-    _ValueType&
-    get_for_current_thread()
-    {
-        std::size_t __i = omp_get_thread_num();
-        if (!__thread_specific_storage[__i])
-        {
-            // create temporary storage on first usage to avoid extra parallel region and unnecessary instantiation
-            __thread_specific_storage[__i] =
-                std::apply([](_Args... __arg_pack) { return std::make_unique<_ValueType>(__arg_pack...); }, __args);
-            __num_elements.fetch_add(1);
-        }
-        return *__thread_specific_storage[__i];
-    }
-
-    std::vector<std::unique_ptr<_ValueType>> __thread_specific_storage;
-    std::atomic_size_t __num_elements;
-    std::tuple<_Args...> __args;
 };
 
-} // namespace __detail
-
+struct __get_thread_num
+{
+    std::size_t
+    operator()()
+    {
+        return omp_get_thread_num();
+    }
+};
 // enumerable thread local storage should only be created from make function
 template <typename _ValueType, typename... Args>
-__detail::__enumerable_thread_local_storage<_ValueType, Args...>
+oneapi::dpl::__utils::__enumerable_thread_local_storage<_ValueType, oneapi::dpl::__omp_backend::__get_num_threads,
+                                                        oneapi::dpl::__omp_backend::__get_thread_num, Args...>
 __make_enumerable_tls(Args&&... __args)
 {
-    return __detail::__enumerable_thread_local_storage<_ValueType, Args...>(std::forward<Args>(__args)...);
+    return oneapi::dpl::__utils::__enumerable_thread_local_storage<
+        _ValueType, oneapi::dpl::__omp_backend::__get_num_threads, oneapi::dpl::__omp_backend::__get_thread_num,
+        Args...>(std::forward<Args>(__args)...);
 }
-
 } // namespace __omp_backend
 } // namespace dpl
 } // namespace oneapi
