+++ include/oneapi/dpl/pstl/parallel_backend_tbb.h
@@ -34,7 +34,6 @@
 #include <tbb/parallel_invoke.h>
 #include <tbb/task_arena.h>
 #include <tbb/tbb_allocator.h>
-#include <tbb/enumerable_thread_specific.h>
 #if TBB_INTERFACE_VERSION > 12000
 #    include <tbb/task.h>
 #endif
@@ -1307,47 +1306,33 @@ __parallel_for_each(oneapi::dpl::__internal::__tbb_backend_tag, _ExecutionPolicy
     tbb::this_task_arena::isolate([&]() { tbb::parallel_for_each(__begin, __end, __f); });
 }
 
-namespace __detail
+struct __get_num_threads
 {
-
-template <typename _ValueType>
-struct __enumerable_thread_local_storage
-{
-    template <typename... _LocalArgs>
-    __enumerable_thread_local_storage(_LocalArgs&&... __args)
-        : __thread_specific_storage(std::forward<_LocalArgs>(__args)...)
-    {
-    }
-
     std::size_t
-    size() const
+    operator()()
     {
-        return __thread_specific_storage.size();
+        return tbb::this_task_arena::max_concurrency();
     }
-
-    _ValueType&
-    get_for_current_thread()
-    {
-        return __thread_specific_storage.local();
-    }
-
-    _ValueType&
-    get_with_id(std::size_t __i)
-    {
-        return __thread_specific_storage.begin()[__i];
-    }
-
-    tbb::enumerable_thread_specific<_ValueType> __thread_specific_storage;
 };
 
-} // namespace __detail
+struct __get_thread_num
+{
+    std::size_t
+    operator()()
+    {
+        return tbb::this_task_arena::current_thread_index();
+    }
+};
 
 // enumerable thread local storage should only be created from make function
 template <typename _ValueType, typename... Args>
-__detail::__enumerable_thread_local_storage<_ValueType>
+oneapi::dpl::__utils::__enumerable_thread_local_storage<_ValueType, oneapi::dpl::__tbb_backend::__get_num_threads,
+                                                        oneapi::dpl::__tbb_backend::__get_thread_num, Args...>
 __make_enumerable_tls(Args&&... __args)
 {
-    return __detail::__enumerable_thread_local_storage<_ValueType>(std::forward<Args>(__args)...);
+    return oneapi::dpl::__utils::__enumerable_thread_local_storage<
+        _ValueType, oneapi::dpl::__tbb_backend::__get_num_threads, oneapi::dpl::__tbb_backend::__get_thread_num,
+        Args...>(std::forward<Args>(__args)...);
 }
 
 } // namespace __tbb_backend
